:recycle: refactor: 테스트 코드 수정
