--- src/test/java/com/part4/team05/sb01otbooteam05/domain/directMessage/service/DirectMessageServiceTest.java
@@ -1,5 +1,0 @@
-package com.part4.team05.sb01otbooteam05.domain.directMessage.service;
-
-public class DirectMessageServiceTest {
-
-}
