--- src/test/java/com/part4/team05/sb01otbooteam05/domain/follow/service/FollowServiceTest.java
@@ -1,5 +1,0 @@
-package com.part4.team05.sb01otbooteam05.domain.follow.service;
-
-public class FollowServiceTest {
-
-}
